[Dataset] [DataFrame 1/n] Refactor table block structure to support potentially more block formats (#20721)
This is the first PR of a series of Ray Dataset and Pandas integration improvements.

This PR refactors `ArrowRow`, `ArrowBlockBuilder`, `ArrowBlockAccessor` by extracting base classes `TableRow`, `TableBlockBuilder`, `TableBlockAccessor`, which can then be inherited for pandas DataFrame support in the next PR.
