+++ python/ray/data/impl/compute.py
@@ -5,7 +5,7 @@ import ray
 from ray.data.block import Block, BlockAccessor, BlockMetadata, \
     BlockPartition, BlockExecStats
 from ray.data.context import DatasetContext
-from ray.data.impl.arrow_block import DelegatingArrowBlockBuilder
+from ray.data.impl.delegating_block_builder import DelegatingBlockBuilder
 from ray.data.impl.block_list import BlockList
 from ray.data.impl.progress_bar import ProgressBar
 from ray.data.impl.remote_fn import cached_remote_fn
@@ -42,7 +42,7 @@ def _map_block_nosplit(block: Block, fn: Any,
                        input_files: List[str]) -> Tuple[Block, BlockMetadata]:
     start_time, start_cpu = time.perf_counter(), time.process_time()
     exec_stats = BlockExecStats()
-    builder = DelegatingArrowBlockBuilder()
+    builder = DelegatingBlockBuilder()
     for new_block in fn(block):
         builder.add_block(new_block)
     new_block = builder.build()
