+++ python/ray/data/read_api.py
@@ -24,8 +24,8 @@ from ray.data.dataset import Dataset
 from ray.data.datasource import Datasource, RangeDatasource, \
     JSONDatasource, CSVDatasource, ParquetDatasource, BinaryDatasource, \
     NumpyDatasource, ReadTask
-from ray.data.impl.arrow_block import ArrowRow, \
-    DelegatingArrowBlockBuilder
+from ray.data.impl.delegating_block_builder import DelegatingBlockBuilder
+from ray.data.impl.arrow_block import ArrowRow
 from ray.data.impl.block_list import BlockList
 from ray.data.impl.lazy_block_list import LazyBlockList
 from ray.data.impl.remote_fn import cached_remote_fn
@@ -58,7 +58,7 @@ def from_items(items: List[Any], *, parallelism: int = 200) -> Dataset[Any]:
     metadata: List[BlockMetadata] = []
     i = 0
     while i < len(items):
-        builder = DelegatingArrowBlockBuilder()
+        builder = DelegatingBlockBuilder()
         for item in items[i:i + block_size]:
             builder.add(item)
         block = builder.build()
