+++ python/ray/data/tests/conftest.py
@@ -5,6 +5,7 @@ import pyarrow as pa
 
 import ray
 
+from ray.data.block import BlockAccessor
 from ray.data.tests.mock_server import *  # noqa
 from ray.data.datasource.file_based_datasource import BlockWritePathProvider
 
@@ -85,7 +86,7 @@ def test_block_write_path_provider():
                                       block=None,
                                       block_index=None,
                                       file_format=None):
-            num_rows = len(ray.get(block)[0])
+            num_rows = BlockAccessor.for_block(ray.get(block)).num_rows()
             suffix = f"{block_index:06}_{num_rows:02}_{dataset_uuid}" \
                      f".test.{file_format}"
             print(f"Writing to: {base_path}/{suffix}")
